+++ mypy/main.py
@@ -500,13 +500,21 @@ def process_options(args: List[str],
         help="Show program's version number and exit",
         stdout=stdout)
 
-    general_group.add_argument(
+    based_group = parser.add_argument_group(
+        title='Based functionality arguments')
+    based_group.add_argument(
         '--write-baseline', action="store_true",
         help="Create an error baseline from the result of this execution")
-    general_group.add_argument(
+    based_group.add_argument(
         '--baseline-file', action='store',
         help="Use baseline info in the given file"
              "(defaults to '{}')".format(defaults.BASELINE_FILE))
+    based_group.add_argument(
+        '--disable-implicit-typing-globals', action="store_false",
+        dest="implicit_typing_globals",
+        help="Disallow using members from `typing` in annotations without imports."
+    )
+
     config_group = parser.add_argument_group(
         title='Config file',
         description="Use a config file instead of command line arguments. "